requirements.txt
django==3.0.1
djangorestframework==3.11.0
aylien-apiclient==0.7.0